Switch back to integer to infer course return type correct

+++ src/db/schema.ts
@@ -5,7 +5,7 @@ import {
   timestamp,
   boolean,
   uuid,
-  numeric,
+  integer,
 } from "drizzle-orm/pg-core";
 
 export const coursesTable = pgTable("courses_table", {
@@ -14,7 +14,7 @@ export const coursesTable = pgTable("courses_table", {
   name: text("name").notNull(),
   description: text("description").notNull(),
   availableForPurchase: boolean("available_for_purchase").notNull(),
-  price: numeric("price").notNull(),
+  price: integer("price").notNull(),
   currency: text("currency").notNull(),
   imagePath: text("image_path"),
   createdAt: timestamp("created_at").notNull().defaultNow(),
@@ -27,7 +27,7 @@ export const purchaseTable = pgTable("purchase_table", {
   id: serial("id").primaryKey(),
   userId: uuid("user_id").notNull(),
   courseId: uuid("course_id").notNull(),
-  amount: numeric("amount").notNull(),
+  amount: integer("amount").notNull(),
   currency: text("currency").notNull(),
   activationId: text("activation_id").notNull(),
   status: text("status").notNull().default("pending"),
